requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
ta-lib>=0.4.25
backtrader>=1.9.76.123
optuna>=3.4.0
pydantic>=2.5.0
websockets>=12.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.17.0
tqdm>=4.66.0
requests>=2.31.0
aiohttp>=3.9.0
apscheduler>=3.10.0
mlflow>=2.8.0  # Optional for model registry
slack-sdk>=3.26.0  # Optional for alerts
discord-webhook>=1.3.0  # Optional for alerts
